Moved minerStart() and minerStop() methods to the JsonRpc2_0Geth class
