+++ geth/src/main/java/org/web3j/protocol/geth/Geth.java
@@ -5,6 +5,7 @@ import org.web3j.protocol.admin.Admin;
 import org.web3j.protocol.admin.methods.response.BooleanResponse;
 import org.web3j.protocol.admin.methods.response.PersonalSign;
 import org.web3j.protocol.core.Request;
+import org.web3j.protocol.core.methods.response.MinerStartResponse;
 import org.web3j.protocol.geth.response.PersonalEcRecover;
 import org.web3j.protocol.geth.response.PersonalImportRawKey;
 
@@ -16,12 +17,16 @@ public interface Geth extends Admin {
         return new JsonRpc2_0Geth(web3jService);
     }
         
-    public Request<?, PersonalImportRawKey> personalImportRawKey(String keydata, String password);
+    Request<?, PersonalImportRawKey> personalImportRawKey(String keydata, String password);
     
-    public Request<?, BooleanResponse> personalLockAccount(String accountId);
+    Request<?, BooleanResponse> personalLockAccount(String accountId);
     
-    public Request<?, PersonalSign> personalSign(String message, String accountId, String password);
+    Request<?, PersonalSign> personalSign(String message, String accountId, String password);
     
-    public Request<?, PersonalEcRecover> personalEcRecover(String message, String signiture);
-    
+    Request<?, PersonalEcRecover> personalEcRecover(String message, String signiture);
+
+    Request<?, MinerStartResponse> minerStart(int threadCount);
+
+    Request<?, BooleanResponse> minerStop();
+
 }
