+++ geth/src/test/java/org/web3j/protocol/geth/RequestTest.java
@@ -52,4 +52,20 @@ public class RequestTest extends RequestTester {
         verifyResult("{\"jsonrpc\":\"2.0\",\"method\":\"personal_ecRecover\",\"params\":[\"0xdeadbeaf\",\"0xa3f20717a250c2b0b729b7e5becbff67fdaef7e0699da4de7ca5895b02a170a12d887fd3b17bfdce3481f10bea41f45ba9f709d39ce8325427b57afcfc994cee1b\"],\"id\":1}");
         //CHECKSTYLE:ON
     }
+
+    @Test
+    public void testMinerStart() throws Exception {
+        web3j.minerStart(4).send();
+
+        verifyResult("{\"jsonrpc\":\"2.0\",\"method\":\"miner_start\","
+                + "\"params\":[4],\"id\":1}");
+    }
+
+    @Test
+    public void testMinerStop() throws Exception {
+        web3j.minerStop().send();
+
+        verifyResult("{\"jsonrpc\":\"2.0\",\"method\":\"miner_stop\","
+                + "\"params\":[],\"id\":1}");
+    }
 }
